+++ pkg/gatewayserver/gatewayserver_test.go
@@ -38,6 +38,7 @@ import (
 	"go.thethings.network/lorawan-stack/pkg/gatewayserver/io"
 	"go.thethings.network/lorawan-stack/pkg/gatewayserver/io/basicstationlns/messages"
 	"go.thethings.network/lorawan-stack/pkg/gatewayserver/io/udp"
+	gsredis "go.thethings.network/lorawan-stack/pkg/gatewayserver/redis"
 	"go.thethings.network/lorawan-stack/pkg/gatewayserver/upstream/mock"
 	"go.thethings.network/lorawan-stack/pkg/rpcclient"
 	"go.thethings.network/lorawan-stack/pkg/rpcmetadata"
@@ -84,6 +85,9 @@ func TestGatewayServer(t *testing.T) {
 		},
 	})
 	c.FrequencyPlans = frequencyplans.NewStore(test.FrequencyPlansFetcher)
+	statsRedisClient, statsFlush := test.NewRedis(t, "gatewayserver_test")
+	defer statsFlush()
+	defer statsRedisClient.Close()
 	config := &gatewayserver.Config{
 		RequireRegisteredGateways: false,
 		MQTT: config.MQTT{
@@ -104,6 +108,10 @@ func TestGatewayServer(t *testing.T) {
 		},
 		BasicStation: gatewayserver.BasicStationConfig{
 			Listen: ":1887",
+		},
+		UpdateConnectionStatsDebounceTime: 0,
+		Stats: &gsredis.GatewayConnectionStatsRegistry{
+			Redis: statsRedisClient,
 		},
 	}
 	gs, err := gatewayserver.New(c, config,
@@ -875,6 +883,10 @@ func TestGatewayServer(t *testing.T) {
 							time.Sleep(timeout)
 						}
 
+						conn, ok := gs.GetConnection(ctx, ids)
+						a.So(ok, should.BeTrue)
+						gs.UpdateConnectionStats(ctx, conn)
+
 						stats, err := statsClient.GetGatewayConnectionStats(statsCtx, &ids)
 						if !a.So(err, should.BeNil) {
 							t.FailNow()
@@ -1080,6 +1092,10 @@ func TestGatewayServer(t *testing.T) {
 						case <-time.After(timeout):
 							t.Fatal("Expected downlink timeout")
 						}
+
+						conn, ok := gs.GetConnection(ctx, ids)
+						a.So(ok, should.BeTrue)
+						gs.UpdateConnectionStats(ctx, conn)
 
 						stats, err := statsClient.GetGatewayConnectionStats(statsCtx, &ids)
 						if !a.So(err, should.BeNil) {