+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Gateway connection stats are stored in Redis (see `--gs.update-connection-stats-debounce-time` option)
+
 ### Deprecated
 
 ### Removed